Move SchemaManager to fuel-indexer-schema
