+++ fuel-indexer/src/database.rs
@@ -5,45 +5,7 @@ use wasmer::Instance;
 use crate::{ffi, IndexerError, IndexerResult};
 use fuel_indexer_database::{queries, IndexerConnection, IndexerConnectionPool};
 use fuel_indexer_lib::manifest::Manifest;
-use fuel_indexer_schema::{
-    db::tables::{Schema, SchemaBuilder},
-    schema_version, FtColumn,
-};
-
-/// Responsible for laying down graph schemas, processes schema upgrades.
-pub struct SchemaManager {
-    pool: IndexerConnectionPool,
-}
-
-impl SchemaManager {
-    pub async fn new(db_conn: impl Into<String>) -> IndexerResult<SchemaManager> {
-        let pool = IndexerConnectionPool::connect(&db_conn.into()).await?;
-
-        Ok(SchemaManager { pool })
-    }
-
-    pub async fn new_schema(&self, name: &str, schema: &str) -> IndexerResult<()> {
-        let mut connection = self.pool.acquire().await?;
-
-        // TODO: Not doing much with version, but might be useful if we
-        //       do graph schema upgrades
-        let version = schema_version(schema);
-
-        if !queries::schema_exists(&mut connection, name, &version).await? {
-            let _db_schema =
-                SchemaBuilder::new(name, &version, self.pool.database_type())
-                    .build(schema)
-                    .commit_metadata(&mut connection)
-                    .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn load_schema_wasm(&self, name: &str) -> IndexerResult<Schema> {
-        // TODO: might be nice to cache this data in server?
-        Ok(Schema::load_from_db(&self.pool, name).await?)
-    }
-}
+use fuel_indexer_schema::FtColumn;
 
 /// Database for an executor instance, with schema info.
 #[derive(Debug)]
